Corrected my mixup between Lesson3.cs and Checks.cs methods
