+++ Src/BootCamp.Chapter/Checks.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Globalization;
 using System.Text;
 
 namespace BootCamp.Chapter
@@ -14,24 +15,53 @@ namespace BootCamp.Chapter
     /// </summary>
     public static class Checks
     {
+        public static float PromptFloat(string message)
+        {
+            Console.WriteLine(message);
+            float number = 0.0f;
+            bool convertedSuccessfully = false;
+            while (!convertedSuccessfully)
+            {
+                convertedSuccessfully = float.TryParse(
+                    Console.ReadLine(),
+                    NumberStyles.AllowDecimalPoint,
+                    CultureInfo.InvariantCulture,
+                    out number
+                );
+                if (!convertedSuccessfully)
+                {
+                    Console.WriteLine("Please enter a number.");
+                }
+            }
+            return number;
+        }
+
         public static int PromptInt(string message)
         {
-            return Lesson3.PromptInt(message);
+            Console.WriteLine(message);
+            int number = 0;
+            bool convertedSuccessfully = false;
+            while (!convertedSuccessfully)
+            {
+                convertedSuccessfully = Int32.TryParse(Console.ReadLine(), out number);
+                if (!convertedSuccessfully)
+                {
+                    Console.WriteLine("Please enter a number.");
+                }
+            }
+            return number;
         }
 
         public static string PromptString(string message)
         {
-            return Lesson3.PromptString(message);
+            Console.WriteLine(message);
+            return Console.ReadLine();
         }
 
-        public static float PromptFloat(string message)
+        public static float CalculateBmi(float weightKg, float heightM)
         {
-            return Lesson3.PromptFloat(message);
-        }
+            return weightKg / (heightM * heightM);
 
-        public static float CalculateBmi(float weight, float height)
-        {
-            return Lesson3.CalculateBmi(weight, height);
         }
     }
 }
